[디노] step6-1 스프링 카페 7단계 - ajax 댓글 구현 (#62)
* Refactor: userId말고 name을 보여주도록 함
dto의 필드도 name으로 변경

* Feat: jquery 최신 버전으로 변경

* Feat: RestController로 변경하고, ajax 사용
컨트롤러의 반환 json으로 변경하고, show에서 보내는 요청도 ajax로 변경
Dto를 생성해서, 컨트롤러에서 Dto 반환

* Refactor: show에서 alert 안띄우게 변경하고, 로그인 시 메인화면으로 이동

* Feat: Rest컨트롤러를 위한 예외 처리 advice 추가

* Feat: error시 에러 메시지 보여주는 기능 추가

* Feat: delete 기능추가
RestController 안에서 예외가 발생해도 일반 ControllerAdvice가 예외를 가져가는 일 발생
RestControllerAdvice에 우선순위를 추가하고, ReplyController에 일어나는 예외만 가져가도록 함.

* Fix: ajax로 댓글을 지우거나 추가하면 즉시 댓글 수도 변경하도록 수정

* Feat: 로그인 확인 로직 추가
인터셉터로 처리하면 인터셉터 내부의 로직이 너무 복잡해져, 단순하게 컨트롤러에 일단 추가

* Fix: ajax로 추가한 댓글이 줄바꿈 안되는 문제 해결

* Feat: db h2에서 mysql로 변경

* Feat: 서버에서 UTC 시간을 전해주고 클라이언트에서 브라우저의 시간을 쓰도록 변경
서울로 시간을 고정하지 않고, js를 통해 브라우저의 시간에 맞춰 시간을 보여주도록 변경

* Feat: 배포용 properties 추가

* Docs: 배포 URL 추가

* Docs: README 작성

* Feat: 설정파일에서 유저네임과 비밀번호 환경변수로 뺌

+++ src/main/java/codesquad/codestagram/controller/ReplyController.java
@@ -2,15 +2,14 @@ package codesquad.codestagram.controller;
 
 import codesquad.codestagram.domain.Reply;
 import codesquad.codestagram.domain.User;
+import codesquad.codestagram.dto.ReplyResponseDto;
+import codesquad.codestagram.exception.NotLoggedInException;
 import codesquad.codestagram.service.ReplyService;
 import jakarta.servlet.http.HttpSession;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
 
-@Controller
+@RestController
 public class ReplyController {
 
     private final ReplyService replyService;
@@ -20,17 +19,25 @@ public class ReplyController {
     }
 
     @PostMapping("/articles/{index}/answers")
-    public String addAnswer(@PathVariable Long index, @RequestParam String text, HttpSession session) {
+    public ResponseEntity<ReplyResponseDto> addAnswer(@PathVariable Long index, @RequestParam String text, HttpSession session) {
         User user = (User) session.getAttribute("loginUser");
-        replyService.addReplyToArticle(user, index, text);
-        return "redirect:/articles/{index}";
+        if (user == null) {
+            throw new NotLoggedInException("로그인이 필요합니다.");
+        }
+
+        Reply reply = replyService.addReplyToArticle(user, index, text);
+        return ResponseEntity.ok(ReplyResponseDto.from(reply));
     }
 
     @DeleteMapping("/articles/{articleId}/answers/{replyId}")
-    public String addAnswer(@PathVariable Long articleId, @PathVariable Long replyId, HttpSession session) {
+    public ResponseEntity<?> addAnswer(@PathVariable Long articleId, @PathVariable Long replyId, HttpSession session) {
         User user = (User) session.getAttribute("loginUser");
+        if (user == null) {
+            throw new NotLoggedInException("로그인이 필요합니다.");
+        }
+
         Reply reply = replyService.findReplyIfOwner(user,replyId);
         replyService.deleteReply(reply);
-        return "redirect:/articles/" + articleId;
+        return ResponseEntity.noContent().build();
     }
 }
